+++ e2e_tests/browser/basic.py
@@ -7,15 +7,24 @@ from time import sleep
 import os
 import sys
 
+def assert_running(process):
+    try:
+        code = process.wait(1)
+        raise Exception("Server died with return code %d" % code)
+    except subprocess.TimeoutExpired:
+        pass
+
 def eprint(msg):
     print(msg, file=sys.stderr)
 
 port = 4242
 
-server = subprocess.Popen(["cargo", "run", "--features=mock_system", "--", "--bind-port", str(port), "--pg-uri=x"])
+server = subprocess.Popen(["cargo", "run", "--manifest-path", "selfhost-dashboard/Cargo.toml", "--features=mock_system", "--", "--bind-port", str(port), "--pg-uri=x"])
 
 try:
     sleep(3)
+    # Check that the server started
+    assert_running(server)
 
     uri = "http://localhost:" + str(port) + "/dashboard"
 
@@ -41,5 +50,10 @@ try:
     driver.find_element_by_name("password").send_keys("123")
     driver.find_element_by_name("password").send_keys(Keys.RETURN)
 
+    sleep(1)
+    # Check that the server didn't die during test
+    assert_running(server)
+
 finally:
-    server.kill()
+    if server.returncode is None:
+        server.kill()
